Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

This might be considered a matter of personal taste, but I think its nicer with the option to configure the notifications through the CM Settings. The hero has both a jogball and a green/amber speaker led and users have different preference on which each notification should show.

Patch-set: 2