Update patch set 1
Patch Set 1:

As it's not recomended to use journaling in some/most SD  cards, I prefer to use ext2. Most of the time, this additionnal partition will be formatted in ext2?

Why don't you use 'auto' instead of 'ext3'? (If auto is supported by mount cmd)

By the way, I think it is alway a good idea to have acces to all availables partitions in recovery!

Patch-set: 1